+++ tests/testthat/test-roxytest-testexamples-show_sig_profile.R
@@ -2,7 +2,7 @@
 
 # File R/show_sig_profile.R: @testexamples
 
-test_that("Function show_sig_profile() @ L104", {
+test_that("Function show_sig_profile() @ L105", {
   
   # Load SBS signature
   load(system.file("extdata", "toy_mutational_signature.RData",